requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blog==1.0.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
django-apscheduler==0.6.2
django-email-users==2.0.0
docopt==0.6.2
executing==2.0.1
idna==3.7
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.18.0
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2rc1
redis==5.0.6
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
